Move activity notification tests to mail plugin.
